feat: ToolboxMosaico panel + cursor-aware query assist on core 0.5.x (#106)
* feat: ToolboxMosaico panel + host plumbing on core 0.5.x

Squashed Mosaico-port work for PJ4, rebased onto latest main.

- pj_runtime: ToolboxRuntimeHost + QSettingsBackend (toolbox-launch foundation)
- pj_app: Mosaico toolbox panel launcher + CredentialsEditor wiring
- pj_dialog_host: opt-in caret-aware code editor; table in-place updates +
  visible_rows filtering; tree-like draggable table header (Interactive +
  fill-first, hover dividers); named-icon resolver; CredentialsEditor
  registered in PjUiLoader (fixes the empty cert sub-dialog); panel-engine
  idle early-out
- pj_widgets: date-only DateRangePicker (time selectors removed)
- resources: Material named icons (contract / plug_connect / calendar_month /
  refresh); Mosaico panel bordered column band + table-header top border
- core submodule -> 0.5.1 (code-editor caret protocol + opt-in caret tracking)

Rebased onto main (#61 streaming, #93 rounded icons + input chrome, #95
clip-window playback); merge auto-resolved.


* fix(pj_scene2D): decode canonical pj_image_v1 image blobs (Bayer + PNG-wrapped raw)

Toolbox image topics push serialized sdk::Image blobs (image_codec=pj_image_v1)
with no MessageParser. The re-platform onto main dropped the consumer decode
path, so Media2DDockWidget fell back to the JPEG pipeline and rendered black.

- ImagePipelineSource: restore the CanonicalImageCodec route (deserialize each
  blob per frame) and funnel it plus the parser route through one
  decodeCanonicalImage() path.
- Add Bayer demosaic (BayerDecode) and recover raw/bayer pixel buffers wrapped
  losslessly in a grayscale PNG before reinterpreting at the logical geometry.
- Media2DDockWidget: route topics whose metadata declares image_codec=pj_image_v1
  to the canonical-codec source instead of the JPEG fallback.
- Tests for the canonical, Bayer, and grayscale-PNG-wrapped paths; docs updated.


* style(pj_dialog_host): white code editors + wrapping pane in the light theme

Lua/Python code editors use a light-theme syntax highlighter (dark-on-white
tokens), so the editor and its immediate pane (assist dropdowns + labels) stay
white in the light theme; the dark theme keeps the normal background. The editor
and pane are tagged with the _pj_code_editor / _pj_code_editor_pane dynamic
properties (WA_StyledBackground on the pane) and styled via QSS attribute
selectors. Comments in the QSS body use C-style /* */ — // is invalid Qt QSS and
silently discards the whole stylesheet.


* fix: address PR #106 review findings + restore date/time picker

Review fixes layered on top of the ToolboxMosaico port (gor/mosaico-port).

pj_widgets/CredentialsEditor — domain-neutral:
- Drop the hardcoded Mosaico `msco_` api-key regex. Default validity is now
  "non-empty"; a plugin supplies its own format via the new `apiKeyPattern`
  Q_PROPERTY declared in its .ui (applied by QUiLoader). No SDK change.

pj_app/MainWindow toolbox host:
- onCloudToolboxRequested: route every failure (not-found/bind/dialog/
  openPanel/present) through diagnostic_history_ instead of stderr-only
  qWarning, so a failed cloud launch is visible in the UI.
- Delete the panel widget that leaked on a presentPanel reject (engine holds
  only a non-owning QPointer).
- Swap the central panel via QSplitter::replaceWidget so the pane count and
  saved size list stay consistent (was insert+hide → N+1 panes vs N sizes).
- Give PanelSession an explicit ~teardown() so the load-bearing destruction
  order survives a future member reorder.

pj_scene2D image pipeline:
- Use the full 8-byte PNG signature and fall back to raw when a sniffed
  container fails to decode (was: drop the frame), keeping JPEG support.
- Stop calling JPEG wrapping "lossless" in the code comments and
  datatypes_2D.md (PNG is lossless, JPEG is lossy/best-effort); mark `webp`
  reserved (no built-in decoder).
- BayerDecode: pass the measured channel through exactly instead of averaging
  it with neighbours (matches the documented contract; sharper output).
  Add a test that the too-small-buffer guard rejects rather than reads OOB.

pj_app/LeftPanel + pj_scene2D/Media2DDockWidget:
- Warn on malformed manifest/metadata JSON instead of silently dropping the
  cloud toolbox / image topic.

pj_dialog_host: add <algorithm> to widget_binding.cpp; fix the openPanel
ownership example in panel_engine.hpp.

pj_widgets/DateRangePicker: restore TimePickerWidget (time-of-day selection),
reverting the PR's date-only change, per request.

Build green; full test suite 102/102.


* fix(pj_app): use loop variable tb (not toolbox) in LeftPanel manifest warning

Fixes a compile error introduced in e2d41e1: the cloud-toolbox loop binds
'tb', but the malformed-manifest qWarning referenced 'toolbox'.


---------

+++ pj_widgets/include/pj_widgets/SvgUtil.h
@@ -46,11 +46,11 @@ inline void RecolorSvgInk(QByteArray& svg_data, bool light_theme) {
 
   // (3) Root-tag fill injection: if the SVG has no `fill` on its root
   // element, give it one so any per-path-fill-less children inherit it.
-  const int svg_open = svg_data.indexOf("<svg");
+  const int svg_open = static_cast<int>(svg_data.indexOf("<svg"));
   if (svg_open < 0) {
     return;
   }
-  const int tag_end = svg_data.indexOf('>', svg_open);
+  const int tag_end = static_cast<int>(svg_data.indexOf('>', svg_open));
   if (tag_end <= svg_open) {
     return;
   }
